Updated the UserCOntroller to use the UserLoginRequest
